improved structure and minor performance enhancements. usage example in ts
