+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-scrollsy",
   "private": false,
-  "version": "1.1.7",
+  "version": "1.1.8",
   "description": "",
   "license": "MIT",
   "type": "module",
