+++ src/functions/childrenAsMethod.ts
