+++ src/types.ts
@@ -10,6 +10,10 @@ type TScrollObject = {
   end: number;
 }
 
+export type TTimeout = number | null;
+
+export type TScrollingElement = HTMLElement | null | undefined;
+
 export type TChildren = ((props: IScrollDataChildren | IScrollObject) => unknown)
   | {
     children: React.ReactNode
